Module google_cloudprivatecatalog1_beta1::api

Structs

CloudPrivateCatalog
Central instance to access all CloudPrivateCatalog related resource activities

FolderCatalogSearchCall
Search Catalog resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

FolderMethods
A builder providing access to all methods supported on folder resources. It is not used directly, but through the CloudPrivateCatalog hub.

FolderProductSearchCall
Search Product resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

FolderVersionSearchCall
Search Version resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

GoogleCloudPrivatecatalogV1beta1Catalog
The readonly representation of a catalog computed with a given resource context.

GoogleCloudPrivatecatalogV1beta1Product
The readonly representation of a product computed with a given resource context.

GoogleCloudPrivatecatalogV1beta1SearchCatalogsResponse
Response message for PrivateCatalog.SearchCatalogs.

GoogleCloudPrivatecatalogV1beta1SearchProductsResponse
Response message for PrivateCatalog.SearchProducts.

GoogleCloudPrivatecatalogV1beta1SearchVersionsResponse
Response message for PrivateCatalog.SearchVersions.

GoogleCloudPrivatecatalogV1beta1Version
The consumer representation of a version which is a child resource under a Product with asset data.

OrganizationCatalogSearchCall
Search Catalog resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the CloudPrivateCatalog hub.

OrganizationProductSearchCall
Search Product resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

OrganizationVersionSearchCall
Search Version resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

ProjectCatalogSearchCall
Search Catalog resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudPrivateCatalog hub.

ProjectProductSearchCall
Search Product resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

ProjectVersionSearchCall
Search Version resources that consumers have access to, within the scope of the consumer cloud resource hierarchy context.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.